+++ packages/docs/docs/main/change-log.md
@@ -11,6 +11,7 @@ This page lists the changes to the IsaacScript framework.
 
 ## August 18th, 2022 (Unreleased)
 
+- The `Card` enum is now renamed to `CardType` in order to be more consistent with the `CollectibleType` and the `TrinketType` enums. Many helper functions have also changed accordingly, like `isVanillaCard` --> `isVanillaCardType`.
 - Added the following helper functions:
   - `removeAllActiveItems`
   - `removeAllPlayerTrinkets`